# docs/openapi.yaml
openapi: 3.0.3
info:
  title: klin API (kiln-api)
  version: "1.0.0"
  description: |
    REST + SSE backend for klin. Auth is a Clerk bearer token. All routes except
    `/api/auth/*`, `/api/share/*`, `/api/docs` and `/healthz` require a signed-in
    session with org membership; handlers explicitly scope service-role queries.
servers:
  - url: https://kiln-api-2lhz.onrender.com
    description: Production (Render)
  - url: http://localhost:8787
    description: Local dev
...
components:
  securitySchemes:
    clerkBearer:
      type: http
      scheme: bearer
...
security:
  - clerkBearer: []
paths:
...
  /api/connections/{id}/connect:
    post:
      summary: Validate and connect a provider token (GitHub uses OAuth)
      tags: [connections]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, writeOnly: true, minLength: 8, maxLength: 4096 }
      responses:
        '201': { description: Connected after provider validation }
        '400': { description: Invalid connector or OAuth provider }
        '422': { description: Provider rejected the credential }
  /api/connections/{id}:
    delete:
      summary: Disconnect a connector
      tags: [connections]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
...